requirements.txt
# ── API framework ─────────────────────────────────────────────────────────────
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.11.3
python-dotenv==1.1.0

# ── PII detection & redaction ─────────────────────────────────────────────────
presidio-analyzer==2.2.356
presidio-anonymizer==2.2.356
spacy==3.8.4
# Download NLP models after install:
#   python -m spacy download en_core_web_lg          (English — required)
#   python -m spacy download de_core_news_lg          (German)
#   python -m spacy download fr_core_news_lg          (French)
#   python -m spacy download es_core_news_lg          (Spanish)
#   python -m spacy download pt_core_news_lg          (Portuguese / Brazilian)

# ── Language detection ────────────────────────────────────────────────────────
langdetect==1.0.9

# ── File extraction ───────────────────────────────────────────────────────────
pdfplumber==0.11.6
python-docx==1.1.2

# ── Summarization (optional — falls back to extractive if not set) ────────────
anthropic==0.51.0

# ── MCP server (Option 2 — SSE transport, separate process) ──────────────────
mcp[cli]==1.27.0

# ── Rate limiting & quotas ────────────────────────────────────────────────────
slowapi==0.1.9
# redis==5.2.1  # Uncomment for multi-pod Redis-backed rate limiting

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.3.5
pytest-cov==6.1.0
pytest-mock==3.14.0
pytest-asyncio==0.25.3
httpx==0.28.1

# ── Code quality ──────────────────────────────────────────────────────────────
ruff==0.11.4
black==25.1.0
isort==6.0.1
